+++ R/extract_alleles.R
@@ -1,3 +1,90 @@
+#' Split an HLA typing string into alleles
+#'
+#' @description Takes in a space-separated HLA typing string and splits it into
+#'   its constituent loci and alleles ("A_1", "A_2", "DRB1_1").
+#'
+#'   `extract_alleles_str()` takes in a single string, and returns a named
+#'   character vector of alleles.
+#'
+#'   `extract_alleles_df()` takes in a data frame, where one column contains the
+#'   typing string, and returns the same data frame along with a new column for
+#'   each allele.
+#'
+#' @param string String, space-separated HLA typing.
+#' @param df A data frame.
+#' @param col_typing The name of the column in `df` that contains a
+#'   space-separated HLA typing `string` for each row.
+#' @param loci A string or character vector with the loci you are interested in.
+#'   Only these alleles will be returned. Defaults to all. `DRB.` is used for
+#'   DRB3, DRB4, and DRB5.
+#'
+#' @return Either a character vector or a data frame with the named alleles.
+#' @export
+#'
+#' @examples
+#' typing <- "A1 A2 B7 B8 Cw3 DQ5 DQ8 DR4 DR11 DR52 DR53"
+#' extract_alleles_str(typing, loci = "A")
+#' extract_alleles_str(typing)
+#'
+#' df <- tidyr::tibble(typing = typing)
+#' extract_alleles_df(df, "typing", loci = c("A", "B", "C"))
+#'
+#' # Can also handle newer nomenclature
+#' extract_alleles_str("DQB1*03:01 DQB1*05:01 DRB1*04:AMR",
+#'                     loci = c("DRB1", "DQB1"))
+
+extract_alleles_str <- function(
+    string,
+    loci = c("A", "B", "C", "DPB1", "DQA1", "DQB1", "DRB1", "DRB.")) {
+  loci <- rlang::arg_match(loci, multiple = TRUE)
+  # TODO: document
+
+  # get rid of any leading/trailing/double spaces
+  string <- stringr::str_squish(string)
+
+  extract_str <- function(locus, string) {
+    pattern <- build_pattern(locus)
+    allele_names <- stringr::str_c(locus, c("_1", "_2"))
+
+    # return named list (locus_allele) with matches
+    rlang::set_names(
+      stringr::str_match(string, pattern)[, 2:3],
+      allele_names
+    )
+  }
+
+  purrr::map(loci, \(x) extract_str(x, string)) |>
+    purrr::flatten_chr()
+}
+
+#' @rdname extract_alleles_str
+#' @export
+extract_alleles_df <- function(
+    df,
+    col_typing,
+    loci = c("A", "B", "C", "DPB1", "DQA1", "DQB1", "DRB1", "DRB.")) {
+  loci <- rlang::arg_match(loci, multiple = TRUE)
+  # TODO: document
+
+  # get rid of any leading/trailing/double spaces
+  df[col_typing] <- stringr::str_squish(df[col_typing])
+
+  extract_df <- function(locus, df, col_typing) {
+    pattern <- build_pattern(locus)
+    allele_names <- stringr::str_c(locus, c("_1", "_2"))
+
+    tidyr::extract(df, tidyr::all_of(col_typing),
+      into = allele_names,
+      regex = pattern,
+      remove = FALSE
+    )
+  }
+
+  # Split typing for each locus into two columns; combine resulting data frames
+  purrr::map(loci, \(x) extract_df(x, df, col_typing)) |>
+    purrr::reduce(dplyr::full_join)
+}
+
 build_pattern <- function(locus) {
   regexps <- list(
     # don't match if locus is preceded by : or other capital letter. This
@@ -30,54 +117,4 @@ build_pattern <- function(locus) {
     stringr::str_glue(regexps$allele, locus = regexps$loci[locus]),
     "?" # second allele is optional
   )
-}
-
-extract_alleles_df <- function(
-    df,
-    col_typing,
-    loci = c("A", "B", "C", "DPB1", "DQA1", "DQB1", "DRB1", "DRB.")) {
-  loci <- rlang::arg_match(loci, multiple = TRUE)
-  # TODO: document
-
-  # get rid of any leading/trailing/double spaces
-  df[col_typing] <- stringr::str_squish(df[col_typing])
-
-  extract_df <- function(locus, df, col_typing) {
-    pattern <- build_pattern(locus)
-    allele_names <- stringr::str_c(locus, c("_1", "_2"))
-
-    tidyr::extract(df, tidyr::all_of(col_typing),
-      into = allele_names,
-      regex = pattern,
-      remove = FALSE
-    )
-  }
-
-  # Split typing for each locus into two columns; combine resulting data frames
-  purrr::map(loci, \(x) extract_df(x, df, col_typing)) |>
-    purrr::reduce(dplyr::full_join)
-}
-
-extract_alleles_str <- function(
-    string,
-    loci = c("A", "B", "C", "DPB1", "DQA1", "DQB1", "DRB1", "DRB.")) {
-  loci <- rlang::arg_match(loci, multiple = TRUE)
-  # TODO: document
-
-  # get rid of any leading/trailing/double spaces
-  string <- stringr::str_squish(string)
-
-  extract_str <- function(locus, string) {
-    pattern <- build_pattern(locus)
-    allele_names <- stringr::str_c(locus, c("_1", "_2"))
-
-    # return named list (locus_allele) with matches
-    rlang::set_names(
-      stringr::str_match(string, pattern)[, 2:3],
-      allele_names
-    )
-  }
-
-  purrr::map(loci, \(x) extract_str(x, string)) |>
-    purrr::flatten_chr()
 }