+++ NEWS.md
@@ -1,3 +1,8 @@
+# hlapro (development version)
+
+* New `extract_alleles()` gets alleles for each locus from HLA typing string,
+  and separates them into new columns in a data frame.
+
 # hlapro 0.1.0
 
 * New `get_mismatches()` determines mismatched antigens from a pair of