+++ NEWS.md
@@ -1,7 +1,8 @@
 # hlapro (development version)
 
-* New `extract_alleles()` gets alleles for each locus from HLA typing string,
-  and separates them into new columns in a data frame.
+* New `extract_alleles_str()`, `extract_alleles_df()` gets alleles for each 
+  locus from HLA typing string, and separates them into named list elements 
+  / new columns in a data frame.
 
 # hlapro 0.1.0